auth-system-verification-through-mail/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Mail
Flask-WTF
Werkzeug
Flask-Bcrypt
Flask-Login
psycopg2
itsdangerous
python-dotenv
